+++ glue/edge.py
@@ -3,7 +3,7 @@ import os
 from shutil import which
 from typing import Optional
 
-from glue.chromium import is_windows, run  # shared Chromium launcher API
+from glue.browsers_launcher import find_app_path_win, is_windows, run
 
 name: str = 'Microsoft Edge'
 
@@ -15,18 +15,9 @@ def find_path() -> Optional[str]:
 
 
 def _find_edge_win() -> Optional[str]:
-    import winreg as reg
-
-    reg_path = r'SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths\msedge.exe'
-    for install_type in reg.HKEY_CURRENT_USER, reg.HKEY_LOCAL_MACHINE:
-        try:
-            reg_key = reg.OpenKey(install_type, reg_path, 0, reg.KEY_READ)
-            edge_path = reg.QueryValue(reg_key, None)
-            reg_key.Close()
-            if edge_path and os.path.isfile(edge_path):
-                return edge_path
-        except OSError:
-            pass
+    path = find_app_path_win('msedge.exe')
+    if path:
+        return path
 
     candidates = [
         os.path.expandvars(r'%ProgramFiles(x86)%\Microsoft\Edge\Application\msedge.exe'),